sagemaker/requirements.txt
# Python dependencies for YOLOv8 training on SageMaker
# Based on ultralytics YOLOv8 requirements

# Core ML libraries
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0

# Image processing
opencv-python>=4.8.0
Pillow>=9.5.0
numpy>=1.24.0

# Data handling
pandas>=2.0.0
pyyaml>=6.0

# Utilities
tqdm>=4.65.0
psutil>=5.9.0

# SageMaker SDK (for local testing)
sagemaker>=2.0.0

# Additional dependencies for YOLOv8
scipy>=1.10.0
matplotlib>=3.7.0
seaborn>=0.12.0

